apps/forAUTH/urls.py
from django.urls import path
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenVerifyView,
)

from apps.forAUTH.views import (
    UserDetailView,
    UserLogoutView,
    UserRegisterView,
    ChangePasswordView,
    ResetPasswordView,
)

urlpatterns = [
    path("api/login/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    path("api/token/verify/", TokenVerifyView.as_view(), name="token_verify"),
    path("register/", UserRegisterView.as_view(), name="register"),
    path("logout/", UserLogoutView.as_view(), name="logout"),
    path("me/", UserDetailView.as_view(), name="me"),
    path("change-password/", ChangePasswordView.as_view(), name="change_password"),
    path(
        "reset-password/<int:pk>/", ResetPasswordView.as_view(), name="reset_password"
    ),
]
