requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.3.0
click==8.1.7
Django==5.0.3
django-jazzmin==2.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.39
pillow==10.2.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing==3.7.4.3
uritemplate==4.1.1
whitenoise==6.7.0
